improve text legibility on ScreenGrooveStats login
Conditionally diffuse miso text to black when RainbowMode is in effect.

Change "Press &START; To Continue" text to use "Common Bold" similar to
ScreenEvaluationCasual.  Short of redesigning this screen for different
UX, I think this is the text we should be making large and bold, since
I've seen casual players get confused and walk away from the machine at
this screen.

If the engine's current language is set to Japanese, use "Common Normal"
instead of "Common Bold."  Update ScreenEvaluation's Casual Mode help
text to do the same.  We'll need a more generalized way to handle fonts
like this if Simply Love supports Korean and/or Chinese in the future,
but one-off conditional checks like this suffice for now...