visually distinguish SortMenu choices from folders
This makes ScreenSelectMusic's SortMenu's choices (like "sort by title",
"sort by artist", "change style to double") visually distinct from
folders-of-choices (like "Sorts", "Options").

Folders-of-choices now:
  * have a background quad
  * have left-aligned text
  * include a folder icon

I'd like to follow-up with a 2nd commit that changes the UX of the
SortWheel to behave more like the MusicWheel:  when a user chooses a
folder-of-choices, the row for that folder will remain visible with its
children rows now visible below it.  The current behavior to to fully
replace all rows when entering/leaving a folder.